+++ packages/core/core.module.ts
@@ -1,19 +1,25 @@
 import {ModuleWithProviders, NgModule} from '@angular/core';
-import {AgmMap} from './directives/map';
+
+import {AgmBicyclingLayer} from './directives/bicycling-layer';
 import {AgmCircle} from './directives/circle';
-import {AgmRectangle} from './directives/rectangle';
+import {AgmDataLayer} from './directives/data-layer';
+import {AgmFitBounds} from './directives/fit-bounds';
 import {AgmInfoWindow} from './directives/info-window';
+import {AgmKmlLayer} from './directives/kml-layer';
+import {AgmMap} from './directives/map';
 import {AgmMarker} from './directives/marker';
 import {AgmPolygon} from './directives/polygon';
-import {AgmPolyline} from './directives/polyline';
+import {AgmPolylineIcon} from './directives/polyline-icon';
 import {AgmPolylinePoint} from './directives/polyline-point';
-import {AgmKmlLayer} from './directives/kml-layer';
-import {AgmDataLayer} from './directives/data-layer';
+import {AgmPolyline} from './directives/polyline';
+import {AgmRectangle} from './directives/rectangle';
+import {AgmTransitLayer} from './directives/transit-layer';
+
 import {LazyMapsAPILoader} from './services/maps-api-loader/lazy-maps-api-loader';
 import {LAZY_MAPS_API_CONFIG, LazyMapsAPILoaderConfigLiteral} from './services/maps-api-loader/lazy-maps-api-loader';
 import {MapsAPILoader} from './services/maps-api-loader/maps-api-loader';
+
 import {BROWSER_GLOBALS_PROVIDERS} from './utils/browser-globals';
-import {AgmFitBounds} from '@agm/core/directives/fit-bounds';
 
 /**
  * @internal
@@ -22,7 +28,7 @@ export function coreDirectives() {
   return [
     AgmMap, AgmMarker, AgmInfoWindow, AgmCircle, AgmRectangle,
     AgmPolygon, AgmPolyline, AgmPolylinePoint, AgmKmlLayer,
-    AgmDataLayer, AgmFitBounds
+    AgmDataLayer, AgmFitBounds, AgmPolylineIcon, AgmTransitLayer, AgmBicyclingLayer
   ];
 }