+++ packages/js-marker-clusterer/services/managers/cluster-manager.ts
@@ -3,11 +3,9 @@ import {Observable, Observer} from 'rxjs';
 
 import 'js-marker-clusterer';
 
-import {MarkerManager} from '../../../core/services/managers/marker-manager';
-import {GoogleMapsAPIWrapper} from '../../../core/services/google-maps-api-wrapper';
-import {AgmMarker} from '../../../core/directives/marker';
-import {AgmMarkerCluster} from './../../directives/marker-cluster';
+import {AgmMarker, GoogleMapsAPIWrapper, MarkerManager} from '@agm/core';
 import {Marker} from '@agm/core/services/google-maps-types';
+import {AgmMarkerCluster} from '../../directives/marker-cluster';
 import {MarkerClustererInstance, ClusterOptions} from '../google-clusterer-types';
 
 declare var MarkerClusterer: any;